Repackage to clarify publicly available classes
Moved internal classes to a separate package as well as renamed classes from utils package and moved them to models.

If you are a developer, you want to primarly use CompatBridge class and classes in the "model", "nbt" and "utils" packages only. Do not use internals package because you will waste server's performance on modern servers.